+++ frontend/public/locales/en-US/analytics.json
@@ -61,7 +61,10 @@
       "minute1": "1 minute",
       "minute5": "5 minutes",
       "minute10": "10 minutes"
-    }
+    },
+    "costTipTitle": "Cost reminder",
+    "costTipContent": "Please note that opening real-time reports will incur additional fees, as detailed in the documentation: ",
+    "costTipLink": "Cost"
   },
   "analyzes": {
     "title": "Analyses",
