+++ frontend/src/index.scss
@@ -588,6 +588,10 @@ ul.menu-list li.checkbox-item {
   gap: 5px;
 }
 
+.cs-analytics-realtime-tips {
+  margin-top: 4px;
+}
+
 .cs-analytics-realtime-auto-refresh {
   float: right;
 }
